Users can onyl delete their own review - functional. Fix queries so that user profile displays even if user doesnt have any reviews

+++ src/server/routes/reviews.js
@@ -4,10 +4,22 @@ const { isAuthorized } = require('../middlewares');
 
 router.delete('/reviews/:id', isAuthorized, (request, response) => {
   const id = request.params.id;
-  Reviews.deleteById(id)
-  .then(() => {
-    const userId = request.session.user.id;
-    response.redirect(`/users/${userId}`);
+  Reviews.getById(id)
+  .then(review => {
+    if (request.session.user.id !== review.user_id) {
+      const previousPage = request.headers.referer;
+      response.status(403);
+      response.render('not-authorized', {previousPage, warning: 'You can only delete your own posts.'});
+    } else {
+      Reviews.deleteById(id)
+      .then(() => {
+        const userId = request.session.user.id;
+        response.redirect(`/users/${userId}`);
+      })
+      .catch(error => {
+      console.log(error);
+      });
+    }
   })
   .catch(error => {
     console.error(error.message);
